+++ CHANGELOG.md
@@ -46,6 +46,9 @@ current (development)
 - Bugfix: Add unicode 13 support for full width characters.
 - Bugfix: Fix MSVC treating codecvt C++17 deprecated function as an error.
 
+### Build
+- Support using the google test version provided by the package manager.
+
 3.0.0
 -----